reorder notary verification components
- [CORE-438]
