+++ dapps/profile.vue/dbcp.json
@@ -7,7 +7,8 @@
       },
       "entrypoint": "profile.vue.js",
       "files": [
-        "profile.vue.js"
+        "profile.vue.js",
+        "profile.vue.css"
       ],
       "origin": "QmQLRfJ6JoqnnaMytkv9yq563zNWJPJAqK5qd1G1coE1zP",
       "primaryColor": "#ac7517",
