+++ dapps/profile.vue/src/components/utils/labeled-list/labeled-list.scss
