+++ dapps/profile.vue/src/components/verifications/notary/actions/pin/pin.scss
