--- dapps/profile.vue/src/components/verifications/notary/verification/verification.ts
@@ -1,218 +1,0 @@
-/*
-  Copyright (C) 2018-present evan GmbH.
-
-  This program is free software: you can redistribute it and/or modify it
-  under the terms of the GNU Affero General Public License, version 3,
-  as published by the Free Software Foundation.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-  See the GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program. If not, see http://www.gnu.org/licenses/ or
-  write to the Free Software Foundation, Inc., 51 Franklin Street,
-  Fifth Floor, Boston, MA, 02110-1301 USA, or download the license from
-  the following URL: https://evan.network/license/
-
-  You can be released from the requirements of the GNU Affero General Public
-  License by purchasing a commercial license.
-  Buying such a license is mandatory as soon as you use this software or parts
-  of it on other blockchains than evan.network.
-
-  For more information, please contact evan GmbH at this address:
-  https://evan.network/license/
-*/
-
-// vue imports
-import Vue from 'vue';
-import Component, { mixins } from 'vue-class-component';
-import { Prop } from 'vue-property-decorator';
-
-// evan.network imports
-import { EvanComponent } from '@evan.network/ui-vue-core';
-import { FileHandler, } from '@evan.network/ui';
-import * as bcc from '@evan.network/api-blockchain-core';
-import * as dappBrowser from '@evan.network/ui-dapp-browser';
-
-import * as dispatchers from '../../../../dispatchers/registry';
-import { notarySmartAgentAccountId } from '../notary.identifications';
-
-@Component({ })
-export default class IdentNotaryVerificationComponent extends mixins(EvanComponent) {
-  /**
-   * label that should be used to be displayed before the verification topic
-   */
-  @Prop() title;
-
-  /**
-   * address for the verification that should be loaded
-   */
-  @Prop() topic;
-
-  /**
-   * address for the contract / account / multi sig for that the verifcations should be checked
-   */
-  @Prop() address;
-
-  /**
-   * ui status flags
-   */
-  loading = true;
-  accepting = false;
-
-  /**
-   * verification details
-   */
-  verification: bcc.VerificationsResultV2 = null;
-
-  /**
-   * All files that are available for the verifications (e.g.: notary certified documents)
-   */
-  files: Array<any> = [ ];
-
-  /**
-   * listen for dispatcher updates
-   */
-  listeners: Array<Function> = [ ];
-
-  /**
-   * Load verification details.
-   */
-  async created() {
-    this.listeners.push(dispatchers.verificationAcceptDispatcher
-      .watch(async ($event) => {
-        // if dispatcher has finished loading, reload the data
-        if ($event.detail.status === 'finished') {
-          await this.loadVerification();
-          this.accepting = false;
-        }
-      }));
-
-    await this.checkLoading();
-    await this.loadVerification();
-
-    this.loading = false;
-  }
-
-  /**
-   * Clear listeners...
-   */
-  beforeDestroy() {
-    this.listeners.forEach(listener => listener());
-  }
-
-  /**
-   * check if currently a verification gets accepted.
-   */
-  async checkLoading() {
-    const runtime: bcc.Runtime = (<any>this).getRuntime();
-    const instances = await dispatchers.verificationAcceptDispatcher.getInstances(runtime);
-
-    this.accepting = instances.length !== 0;
-  }
-
-  /**
-   * Load verification details for the current topic, ensuring, they are created by the evan ens
-   * root owner.
-   */
-  async loadVerification() {
-    const runtime: bcc.Runtime = (<any>this).getRuntime();
-
-    // reset loaded files
-    this.files = [];
-    // load the verification details
-    // allow only ens root owner and notary smart agent
-    const allowedVerificationAccounts = [
-      // ens root owner
-      '0x4a6723fC5a926FA150bAeAf04bfD673B056Ba83D',
-      // notary smart agent
-      notarySmartAgentAccountId
-    ];
-    const verificationQuery = JSON.parse(JSON.stringify(runtime.verifications.defaultQueryOptions));
-    verificationQuery.validationOptions.issued = bcc.VerificationsStatusV2.Yellow;
-    verificationQuery.validationOptions.parentUntrusted = bcc.VerificationsStatusV2.Green;
-    verificationQuery.validationOptions.selfIssued = bcc.VerificationsStatusV2.Green;
-    verificationQuery.validationOptions.notEnsRootOwner = (verification, pr) => {
-      // trust root verifications issued by root account
-      // subject does not need to be root account as well
-      if (allowedVerificationAccounts.indexOf(verification.details.issuer) !== -1) {
-        return bcc.VerificationsStatusV2.Green;
-      }
-      return bcc.VerificationsStatusV2.Red;
-    };
-    // load nested verifications
-    this.verification = await runtime.verifications.getNestedVerificationsV2(
-      this.address,
-      this.topic,
-      false,
-      verificationQuery,
-    );
-
-    this.verification.verifications
-      .forEach(async (subVerification) => {
-        try {
-          const contentKey = await runtime.profile.getBcContract(
-            'contracts',
-            runtime.web3.utils.soliditySha3(`verifications,${subVerification.details.id},contentKey`)
-          );
-          const hashKey = await runtime.profile.getBcContract(
-            'contracts',
-            runtime.web3.utils.soliditySha3(`verifications,${subVerification.details.id},hashKey`)
-          );
-          const fileHash = JSON.parse(subVerification.details.data);
-
-          if (fileHash !== '0x0000000000000000000000000000000000000000000000000000000000000000') {
-            if (subVerification.details.topic === '/evan/company') {
-              const hashFiles = await (<any>runtime.dfs).get(fileHash);
-              const foundFiles = JSON.parse(hashFiles)
-              for (let file of foundFiles) {
-                file.file = await (<any>runtime.dfs).get(file.file, true);
-                file.size = file.file.length;
-                this.files.push(await FileHandler.fileToContainerFile(file));
-              }
-            } else {
-              const cryptoAlgorithFiles = 'aesBlob'
-              const cryptoAlgorithHashes = 'aesEcb'
-              const encodingEncrypted = 'hex'
-              const encodingUnencryptedHash = 'hex'
-              const hashCryptor = runtime.cryptoProvider.getCryptorByCryptoAlgo('aesEcb')
-              const dencryptedHashBuffer = await hashCryptor.decrypt(
-                Buffer.from(fileHash.substr(2), encodingUnencryptedHash), { key: hashKey })
-
-              const retrieved = await (<any>runtime.dfs).get('0x' + dencryptedHashBuffer.toString('hex'), true);
-              const cryptor = runtime.cryptoProvider.getCryptorByCryptoAlgo(cryptoAlgorithFiles)
-              const decrypted = await cryptor.decrypt(retrieved, { key: contentKey });
-
-              if (decrypted) {
-                for (let file of decrypted) {
-                  file.size = file.file.length;
-                  this.files.push(await FileHandler.fileToContainerFile(file));
-                }
-              }
-            }
-          }
-        } catch (ex) {
-          runtime.logger.log(`Could not decrypt verification files: ${ ex.message }`, 'error');
-        }
-      });
-  }
-
-  /**
-   * Start the accept verification dispatcher for the current account.
-   */
-  acceptVerification() {
-    this.accepting = true;
-
-    // just accept the first verification
-    dispatchers.verificationAcceptDispatcher.start(
-      (<any>this).getRuntime(),
-      {
-        address: this.address,
-        id: this.verification.verifications[0].details.id,
-        topic: this.topic,
-      }
-    );
-  }
-}
