+++ dapps/profile.vue/src/routes.ts
@@ -35,7 +35,7 @@ import * as dappBrowser from '@evan.network/ui-dapp-browser';
 
 import ProfileDetailComponent from './components/profile/profile.vue';
 import ProfileSettingsComponent from './components/settings/settings.vue';
-import VerificationsComponent from './components/verifications/notary/overview/overview.vue';
+import VerificationsComponent from './components/verifications/overview/overview.vue';
 
 // map them to element names, so they can be used within templates
 const routeRegistration: Array<RouteRegistrationInterface> = [
